olisa/fix_global_state
